Initial commit
Added code examples to the repo. 

There are some examples of training notebooks I have used to teach the very basics of programming to undergrads. An example of c++ code written as part of a training course I attended. The code from a small side project I completed looking at the backbone ring structures of natural products and some derived pharmaceuticals. And a copy of my GA, (BMPGA), this is an older copy than the current version as that contains unpublished work still. But should still give an idea of what my code looks like.